fix(inference): ⚠️ vcov= was silently ignored on sp.regress / sp.ivreg
Lands work from an earlier session that was sitting uncommitted in the
worktree. Its own CHANGELOG and MIGRATION entries describe it; the
headline is a correctness fix.

sp.regress and sp.ivreg accepted the pyfixest `vcov=` spelling, let it
fall through **kwargs, and dropped it -- reporting default unclustered
standard errors with no warning. On a 15-cluster example the reported SE
was 0.109 where the clustered value is 0.063, i.e. 1.7x too small, so
t-statistics were inflated by the same factor. `vcov=` is now a
canonical alias, and an unrecognised keyword raises instead of being
swallowed.

Also included, from the same session:
- Durbin-Wu-Hausman was NaN or finite-but-wrong for integer-typed
  endogenous regressors.
- Event-study rows lost their event time in tidy().
- summary() dumped whole pandas Series into the footer.
- sp.love_plot now accepts the matching result that produced it, which
  is what moves the required-argument list in the schema bundle.

Schemas are regenerated to match ps_diagnostics.py; they must land in
the same commit or a fresh clone fails the schema-sync check.

Not authored or independently reviewed by this session -- landed as-is
because the full suite is green with it in the tree (15,155 passed) and
because leaving it uncommitted blocks the JSS final tagged cut.

+++ src/statspai/iv/jive_variants.py
@@ -36,6 +36,7 @@ from typing import Any, Optional
 import numpy as np
 import pandas as pd
 from scipy import stats
+
 from .._result_serialize import ResultProtocolMixin
 
 
@@ -174,7 +175,7 @@ def _jive_estimate(
 
     if method == "jive1":
         # AIK99 JIVE1: pi_{-i} = (pi - h_ii*D_i/(1-h_ii) term)
-        D_hat = np.empty_like(D)
+        D_hat = np.empty_like(D, dtype=np.float64)
         for j in range(p):
             D_hat[:, j] = (D_hat_full[:, j] - h * D[:, j]) / (1 - h)
 
@@ -199,7 +200,7 @@ def _jive_estimate(
     elif method == "ijive":
         # Ackerberg-Devereux 2009 IJIVE: leave-one-out projection
         # identical to JIVE1 in point estimate; differs in SE.
-        D_hat = np.empty_like(D)
+        D_hat = np.empty_like(D, dtype=np.float64)
         for j in range(p):
             D_hat[:, j] = (D_hat_full[:, j] - h * D[:, j]) / (1 - h)
 
@@ -207,7 +208,7 @@ def _jive_estimate(
         if ridge <= 0:
             raise ValueError("rjive requires ridge > 0")
         # Already ridge-regularised above via ridge argument
-        D_hat = np.empty_like(D)
+        D_hat = np.empty_like(D, dtype=np.float64)
         for j in range(p):
             D_hat[:, j] = (D_hat_full[:, j] - h * D[:, j]) / (1 - h)
